Renamed temporary testing crate
In case I forget, the intention is to create some temporary
crates then designate 1 as the primary.
Add the other crates as dependencies to the primary crate,
then build.

Then remove the dependencies, build again.

Run ddt clean

Confirm that the only *.rmeta & *.rlib in target/debug/deps
are that of the primary crate, (apparently ddt does not delete
the *.d files, this might be a bug.)

Why did I call it centre?